Fallback to brocollig internal js dir as a last resort when loading modules
