+++ brocolli/js/react_renderer.js
@@ -1,15 +1,17 @@
 const react = require("react")
 const react_dom = require('react-dom/server')
 
-module.exports = function render(app_module){
-    let App = require(app_module)
-    let app_markup = react_dom.renderToString(react.createElement(App))
-    // Order is important, always generate the app state after rendering the app.
-    let app_state = JSON.stringify(global.app_state)
+module.exports = function render(app_module, data){
+    let module = require(app_module)
+    if (module['default']){
+        App = module['default']
+    }else{
+        App = module
+    }
+    props = data || {}
     return {
-        "markup": `<div id='root'>${app_markup}</div>
-        <div id='data_brocolli_app_state' style='display:none'><!--${app_state}--></div>
-        `
+        "markup": react_dom.renderToString(react.createElement(App, props)),
+        "state": JSON.stringify(global.___brocolli___.state),
+        "props": JSON.stringify(props)
      }
 }
-
