+++ brocolli/node_shims.py
@@ -25,7 +25,6 @@ class Global(STPyV8.JSClass):
         ctxt.eval(require["code"])
         ctxt.require = ctxt.custom_require
         ctxt.render = ctxt.require(renderer)
-        ctxt.eval("global.app_state = {___brocolli___:{___state___:{}}}")
 
     def load_python_module(self, name: str):
         return brocolli.load_python_module(name)
