--- brocolli/react_renderer.js
@@ -1,17 +1,0 @@
-const react = require("react")
-const react_dom = require('react-dom/server')
-
-module.exports = function render(app_module, data){
-    let module = require(app_module)
-    if (module['default']){
-        App = module['default']
-    }else{
-        App = module
-    }
-    props = data || {}
-    return {
-        "markup": react_dom.renderToString(react.createElement(App, props)),
-        "state": JSON.stringify(global.___brocolli___.state),
-        "props": JSON.stringify(props)
-     }
-}
